Currently verifies with klee 62a98729881b4acacc97092196e60a540e8b9cd8 in playback mode
This and the last 3 commits represent changes to the tools source code
to effect playback verification with cliver.

The tools needed to be re-written to remove dependencies on external functions
that allocate memory--as we cannot access that memory in klee.

The client.cc Client function was re-written, and will now only work in CLIVER.
We need to go back and add functionality for the original Client.  Current client
supports a subset of previously supported commandline options.

Connect has also changed both in behavior (to avoid 0length recv behavior after
shutdown), and in function definition.  These changes have not propagated beyond
our immediate needs (in Client).